+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # The Delightful World of Vue
 ## [12.02.2025]
+- Page Context "Service"
 - Reading Server Data & :global Classes
 - Passing data From the Server to Vue
 - Dynamic Categories via Ajax